requirements.txt
streamlit
open-clip-torch
torch
tqdm
pandas
scikit-learn
Pillow
